Forward/Backward Object Detection
It's very buggy, but fundamentally it works. Added rays to the object on
initialization. Each render loop checks the rays for their distance from
an object, then sets the speed in that direction to 0 if the distance is
small enough. These vectors are rotated when the player rotates. Lots of
green code here still, needs cleanup.