Object addition attempt and fixes
Fixed the movement controls, but changed some things in respect to
player. Added a raycastReference object so that we dont accidentally
rotate the raycasters when we roll the ball. Also detached the camera
from the ball and had it follow the player controls instead.

Currently the add object on touch does not work,  it adds to a cylinder
child (used to show what the ball rotation will look like when we get
the add function working). Sadly it adds based on world position. If we
try to reset the position of the added object to the childs position, it
works, except everything stacks in one spot. This will be fixed once we
can get the objects to attach to the child based on its current rotation
as well. Not sure how this will be done.