Changed some camera features and raycast features
The raycasting for object detection is now dynamic. the distance
variable increments slowly as objects are added so that the ball appears
to be growing.

Also added some updates to the initial camera position as well as the
update to the camera as the ball grows so the user can see more, as
otherwise it would obstruct the view. Look in the conditional for where
objects are added to the player.